requirements.txt
itk-elastix>=0.8.0
numpy>=1.19.0
napari>=0.4.6
napari-plugin-engine>=0.1.4
magicgui>=0.2.6
